requirements.txt
# -------------------------------
# Core Deep Learning (CUDA 12.1)
# Install manually via: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
# -------------------------------
torch
torchvision
torchaudio

# Transformers and Training
transformers==4.38.2
accelerate==0.27.2

# Image and Video Processing
opencv-python==4.9.0.80
pillow==10.3.0
imageio==2.34.0
imageio-ffmpeg==0.4.9
scikit-image==0.22.0
matplotlib==3.8.4

# Visualization and Simulation
pygame==2.5.2

# Utilities
numpy==1.24.4
tqdm==4.66.2
scikit-learn==1.3.2


# IDEs and Interactive Tools
jupyter==1.0.0
notebook==7.1.1
ipykernel==6.29.4
spyder==5.5.1

# Optional: For saving logs, outputs, config parsing
pyyaml==6.0.1
